+++ NEWS.md
@@ -1,10 +1,34 @@
-# journalR 0.2.1
+# journalR 0.6.0
 
-cran initial submit responses
+## Improvements
 
-- description update 
+`format_means_df()` 
+
+- now dispatches `fround_x()` helpers for more consistent formatting, especially for Lancet
+
+## Breaking
+
+- renamed `format_means_df()` to `format_metric_cols()`, with arg name change from `central_var` to `var_prefix`
+
+# journalR 0.5.1
+
+cran initial submit harmonization
+
+- description update
 - remove dontrun examples
 - remove globalenv references
+
+# journalR 0.5.0
+
+mag override can now be passed down from user API
+
+# journalR 0.4.0
+
+ added rate-space metric/metric
+
+# journalR 0.3.0
+
+Internal state management improvements.  No changes to user interface.
 
 # journalR 0.2.0